commit report

+++ application/modules/area/views/area_index.php
@@ -11,7 +11,7 @@
 		<th>ลำดับ</th>
 		<th>เขตความรับผิดชอบ</th>
 		<th>จำนวนเขต</th>
-		<th><a href="area/form" class="btn_add" title="เพิ่ม" name="btn_add"></a></th>
+		<th width="90"><a href="area/form" class="btn" title="เพิ่ม" name="btn_add">เพิ่มรายการ</a></th>
 	</tr>
 	<?php foreach($result as $key=>$item): ?>
 	<tr>
@@ -19,8 +19,8 @@
 		<td><?php echo ++$key ?></td>
 		<td><?php echo $item['name'] ?></td>
 		<td><?php echo  $item['total']?></td>
-		<td><a href="area/form/<?php echo $item['id'] ?>" class="btn" title="แก้ไข">แก้ไข</a>
-				  <a href="area/delete/<?php echo $item['id'] ?>" class="btn" title="ลบ">ลบ</a></td>
+		<td><a href="area/form/<?php echo $item['id'] ?>" 	class="btn" title="แก้ไข">แก้ไข</a>
+			<a href="area/delete/<?php echo $item['id'] ?>" class="btn" title="ลบ" onclick="return confirm('<?php echo NOTICE_CONFIRM_DELETE?>')">ลบ</a></td>
 	</tr>
 	<?php endforeach; ?>
 </table>
